Multi-generation functionality working.
Split much of the dove list manipulation methods into a dove_manager class.
